+++ src/components/plots/Header.jsx
@@ -14,7 +14,7 @@ import { useBeforeunload } from 'react-beforeunload';
 import FeedbackButton from '../FeedbackButton';
 import { savePlotConfig } from '../../redux/actions/componentConfig/index';
 import itemRender from '../../utils/renderBreadcrumbLinks';
-import { getFromApiExpectOK } from '../../utils/getFromApiExpectOK';
+import { getFromApiExpectOK } from '../../utils/getDataExpectOK';
 import { LOAD_CONFIG } from '../../redux/actionTypes/componentConfig';
 import { initialPlotConfigStates } from '../../redux/reducers/componentConfig/initialState';
